咨询表完善 用户表完善 诊断表查询

+++ vision-parent/vision-manage/src/main/java/com/vision/controller/cus/CusConsultationController.java
@@ -1,5 +1,7 @@
 package com.vision.controller.cus;
 
+import com.alibaba.druid.util.StringUtils;
+import com.vision.exception.ServiceException;
 import com.vision.pojo.cus.CusConsultation;
 import com.vision.pojo.cus.vo.CusVo;
 import com.vision.service.cus.CusConsultationService;
@@ -24,10 +26,22 @@ public class CusConsultationController {
     /**基于用户/电话及当前页码值条件查询用户信息*/
     @RequestMapping("/getConsultation")
     @ResponseBody
-    public JsonResult getById( CusVo cusVo){
+    public JsonResult getConsultation( CusVo cusVo){
+    	//1.数据合法性验证
+		Integer pageCurrent = cusVo.getPageCurrent();
+		Integer orgId = cusVo.getOrgId();
+		Integer pageSize = cusVo.getPageSize();
+		if(pageCurrent==null||pageCurrent<=0)
+			return JsonResult.build(201, "页码值不正确");
+		if(orgId<0||orgId==null)
+			return JsonResult.build(201, "门店信息不正确");
+		if(pageSize<0||pageSize==null)
+			return JsonResult.build(201, "页码大小不正确");
+		
+		
     	try {
     		PageObject<CusConsultation> pageObject = cusConsultationService.getConsultation(cusVo);
-    		if(!(pageObject.getRecords().size()==0)) {
+    		if(pageObject.getRecords()!=null&&!(pageObject.getRecords().size()==0)) {
         		return JsonResult.oK(pageObject);
         	}
 		} catch (Exception e) {
@@ -42,6 +56,17 @@ public class CusConsultationController {
     @RequestMapping("/addConsultation")
     @ResponseBody
     public JsonResult addConsultation( CusConsultation cusConsultation) {
+    	
+    	//验证数据合法性
+    	if(cusConsultation==null)
+    		return JsonResult.build(201, "对象不能为空");
+    	if(cusConsultation.getOrgId()<0||cusConsultation.getOrgId()==null)
+    		return JsonResult.build(201, "门店orgId不能为空");
+    	if(StringUtils.isEmpty(cusConsultation.getName()))
+    		return JsonResult.build(201, "用户名不能为空");
+    	if(StringUtils.isEmpty(cusConsultation.getTel()))
+    		return JsonResult.build(201, "电话不能为空");
+    	
     	//获取登录用户信息
     	try {
 			Integer row = cusConsultationService.addConsultation(cusConsultation);
@@ -49,8 +74,10 @@ public class CusConsultationController {
 				return JsonResult.oK();
 			}
 		} catch (Exception e) {
+			e.printStackTrace();
 			System.out.println("将CusCustomer类型数据添加到数据库=============错误==================");
 		}
+    	
     	return JsonResult.build(201, "保存失败");
     }
     
@@ -58,12 +85,19 @@ public class CusConsultationController {
 	@RequestMapping("deleteConsultation")
 	@ResponseBody
 	public JsonResult deleteConsultation(Integer id, Integer  orgId) {
+		//验证数据合法性
+		if(orgId==null||orgId<0)
+			return JsonResult.build(201, "请选择一条数据");
+		if(id==null||id<=0)
+			return JsonResult.build(201, "请选择一条数据");
+		
 		try {
 			Integer row = cusConsultationService.deleteConsultation(id, orgId);
 			if(row != 0 || row == null) {
 				return JsonResult.oK();
 			}
 		} catch (Exception e) {
+			e.printStackTrace();
 			System.out.println("基于id删除咨询表信息=============错误==================");
 		}
 		return JsonResult.build(201, "删除失败,数据或已删除");
@@ -73,12 +107,20 @@ public class CusConsultationController {
 	@RequestMapping("getConsultationById")
 	@ResponseBody
 	public JsonResult findObjectById(Integer id, Integer orgId) {
+		
+		//验证数据合法性
+		if(orgId==null||orgId<0)
+			return JsonResult.build(201, "orgId错误");
+		if(id==null||id<=0)
+			return JsonResult.build(201, "id错误");
+		
 		try {
 			CusConsultation cusConsultation = cusConsultationService.getConsultationById(id, orgId);
 			if(cusConsultation != null) {
 				return JsonResult.oK(cusConsultation);
 			}
 		} catch (Exception e) {
+			e.printStackTrace();
 			System.out.println("基于咨询表id,查询相关id所有信息=============错误==================");
 		}
 		return JsonResult.build(201, "修改查询失败");
@@ -89,6 +131,19 @@ public class CusConsultationController {
 	@RequestMapping("updateConsultation")
 	@ResponseBody
 	public JsonResult updateConsultation(CusConsultation cusConsultation) {
+		
+		//验证数据合法性
+    	if(cusConsultation==null)
+    		return JsonResult.build(201, "对象不能为空");
+		if(cusConsultation.getId()<0||cusConsultation.getId()==null)
+			return JsonResult.build(201, "对象id不能为空");
+    	if(cusConsultation.getOrgId()<0||cusConsultation.getOrgId()==null)
+    		return JsonResult.build(201, "门店orgId不能为空");
+    	if(StringUtils.isEmpty(cusConsultation.getName()))
+    		return JsonResult.build(201, "用户名不能为空");
+    	if(StringUtils.isEmpty(cusConsultation.getTel()))
+    		return JsonResult.build(201, "电话不能为空");
+		
 		//获取登录用户信息
 		try {
 			Integer row = cusConsultationService.updateConsultation(cusConsultation);
@@ -96,6 +151,7 @@ public class CusConsultationController {
 				return JsonResult.oK();
 			}
 		} catch (Exception e) {
+			e.printStackTrace();
 			System.out.println("基于咨询表id更改用户信息=============错误==================");
 		}
 		return JsonResult.build(201, "修改保存失败");
